Streamlined naming, fixed one bug
