+++ reem/__init__.py
@@ -1,4 +1,3 @@
 __version__ = '0.1.0'
 __all__ = ['RedisInterface','KeyValueStore','PublishSpace','CallbackSubscriber','SilentSubscriber']
-from .connection import RedisInterface
-from .datatypes import KeyValueStore,PublishSpace,SilentSubscriber,CallbackSubscriber
+from .connection import RedisInterface,KeyValueStore,PublishSpace,SilentSubscriber,CallbackSubscriber
